renames analytics modules and implements minor changes
